# docs/api.yaml
openapi: 3.1.0
info:
  title: ECHO Public Gateway API
  version: 1.0.0
  description: The browser-facing API. Internal service routes and hostnames are intentionally omitted.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /health:
    get:
      security: []
      summary: Gateway liveness
      responses:
        "200": { description: Healthy }
  /journals:
    get:
      summary: List the authenticated user's journal entries
      responses:
        "200": { description: Journal list }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create an encrypted journal entry
      responses:
        "201": { description: Journal created }
        "400": { description: Invalid journal }
  /journals/draft:
    get:
      summary: Get the user's draft
      responses: { "200": { description: Draft or null } }
    put:
      summary: Save the user's encrypted draft
      responses: { "200": { description: Draft saved } }
    delete:
      summary: Delete the user's draft
      responses: { "204": { description: Draft deleted } }
  /journals/{journalId}:
    parameters:
      - $ref: "#/components/parameters/JournalId"
    get:
      summary: Get a journal entry
      responses: { "200": { description: Journal }, "404": { description: Not found } }
    patch:
      summary: Update a journal entry
      responses: { "200": { description: Journal updated } }
    delete:
      summary: Delete a journal entry
      responses: { "204": { description: Journal deleted } }
  /journals/{journalId}/analyze:
    parameters:
      - $ref: "#/components/parameters/JournalId"
    post:
      summary: Request analysis through Analysis and ML services
      responses:
        "201": { description: Analysis completed }
        "403": { description: Verification, account consent, or per-entry consent required }
        "503": { description: ML or another dependency unavailable }
        "504": { description: Dependency timeout }
  /journals/{journalId}/analyses:
    parameters:
      - $ref: "#/components/parameters/JournalId"
    get:
      summary: Get the latest analysis result
      responses: { "200": { description: Latest analysis or null } }
  /moods:
    get:
      summary: List recent mood entries
      responses: { "200": { description: Mood entries } }
    post:
      summary: Record an encrypted mood entry
      responses: { "201": { description: Mood recorded }, "400": { description: Invalid mood } }
  /assessments/phq8:
    post:
      summary: Calculate a non-diagnostic PHQ-8 screening score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  minItems: 8
                  maxItems: 8
                  items: { type: integer, minimum: 0, maximum: 3 }
      responses: { "200": { description: Score, severity, and disclaimer }, "400": { description: Invalid answers } }
  /recommendations:
    post:
      summary: Request a wellbeing/CBT recommendation
      responses: { "200": { description: Recommendation } }
  /buddy/session:
    get:
      summary: Get or start the verified user's Buddy session
      responses: { "200": { description: Buddy session }, "403": { description: Verification required } }
  /buddy/messages:
    post:
      summary: Send a Buddy message
      responses: { "201": { description: Updated session }, "403": { description: Verification required } }
  /buddy/history:
    get:
      summary: List Buddy conversations
      responses: { "200": { description: Conversation history } }
  /grounding/sessions:
    post:
      summary: Record a grounding session
      responses: { "201": { description: Session recorded } }
  /support-resources:
    get:
      security: []
      summary: List verified active support resources
      responses: { "200": { description: Resources } }
  /dashboard:
    get:
      summary: Get the API-composed dashboard
      responses: { "200": { description: Dashboard } }
  /insights/emotions:
    get:
      summary: Get derived emotion patterns
      responses: { "200": { description: Emotion insights } }
  /settings:
    get:
      summary: Get account settings
      responses: { "200": { description: Settings } }
  /settings/profile:
    patch:
      summary: Update profile settings
      responses: { "200": { description: Profile updated } }
  /settings/privacy:
    patch:
      summary: Update privacy settings
      responses: { "200": { description: Privacy settings updated } }
  /settings/notifications:
    patch:
      summary: Update notification settings
      responses: { "200": { description: Notification settings updated } }
  /settings/trusted-contacts:
    post:
      summary: Create a trusted contact
      responses: { "201": { description: Contact created } }
  /settings/trusted-contacts/{contactId}:
    parameters:
      - name: contactId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Update a trusted contact
      responses: { "200": { description: Contact updated } }
    delete:
      summary: Delete a trusted contact
      responses: { "200": { description: Contact deleted } }
  /settings/data-exports:
    post:
      summary: Request an account data export
      responses: { "201": { description: Export requested } }
  /settings/account-deletion:
    post:
      summary: Request account deletion
      responses: { "201": { description: Deletion requested } }
  /settings/account-deletion/{requestId}/cancel:
    parameters:
      - name: requestId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Cancel a pending account deletion request
      responses: { "200": { description: Deletion request cancelled } }
  /onboarding/status:
    get:
      summary: Get onboarding status
      responses: { "200": { description: Onboarding status } }
  /onboarding/consent:
    post:
      summary: Save onboarding consent choices
      responses: { "200": { description: Consent saved } }
  /onboarding/profile:
    post:
      summary: Save onboarding profile data
      responses: { "200": { description: Profile saved } }
  /onboarding/setup:
    post:
      summary: Save onboarding setup choices
      responses: { "200": { description: Setup saved } }
  /onboarding/complete:
    post:
      summary: Complete onboarding
      responses: { "200": { description: Onboarding completed } }
  /verification:
    get:
      summary: Get verification status
      responses: { "200": { description: Verification status } }
  /verification/application:
    put:
      summary: Save a verification application
      responses: { "200": { description: Application saved } }
  /verification/documents/{kind}:
    parameters:
      - name: kind
        in: path
        required: true
        schema: { type: string }
    put:
      summary: Upload a verification document
      responses: { "200": { description: Document uploaded }, "413": { description: Document too large } }
  /verification/submit:
    post:
      summary: Submit a verification application
      responses: { "200": { description: Application submitted } }
  /admin/verifications:
    get:
      summary: List verification applications for an authorized administrator
      responses: { "200": { description: Verification list }, "403": { description: Administrator access required } }
  /admin/verifications/{verificationId}:
    parameters:
      - name: verificationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get verification detail for an authorized administrator
      responses: { "200": { description: Verification detail } }
  /admin/verifications/{verificationId}/claim:
    parameters:
      - name: verificationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Claim a submitted verification for review
      responses: { "200": { description: Review claimed }, "409": { description: Already claimed or reviewed } }
  /admin/verifications/{verificationId}/decision:
    parameters:
      - name: verificationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Record a verification decision
      responses: { "200": { description: Decision recorded }, "409": { description: Already reviewed } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase access token
  parameters:
    JournalId:
      name: journalId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing, invalid, or expired Supabase access token